horizon_calc: Rework update_parent_horizon to be breadth-first #91
Previously, update_parent_horizon would consider each node to update in
isolation. This meant that all common parents' horizons would be recalculated
for each node, resulting in a lot of wasted time.

Instead, work through nodes to update breadth first, this means that we only
consider a common parent of 2 nodes once.

This does not handle the case when nodes to update are on different levels
well; but that appears to be unlikely enough to not warrant the housekeeping to
spot that situation.