polytomy/otop: Draw undeveloped nodes as branches #91
When a node is undeveloped, node.children is empty, but
node.full_children_length is accurate. Unfortunately node.has_child() uses the
former, so fake_leaf_shapes are assumed not to have children and giant leaves
are drawn. Replace all calls to has_child with full_children_length.

In addition, remove the attempt to include the child length as part of the fake
branch length. It was wrong, and didn't scale properly with zoom levels.