otop: Render fake fans with real nodes if available #143
As we force nodes to be fake in polytomy views, we can attempt to render a fan
when the nodes below are actually developed. If that is the case, use the child
node positions to generate the fans. This, coupled with fixing the radius
calculation, stops spurious lines appearing from off-screen nodes (the node and
it's children are off-screen, but the fake-fan isn't).