setup_page: Re-parse window location, don't recycle state object #166
record_url(), as well as adding the current position to the browser history,
also adds a state object into the browser's history. popupstate attempts to
pick this up and use it to restore the page state.

However, since that state object is intended for making a URL, the x/y/w values
are rounded strings, not integers. This causes the restore to fly off in
arbitrary directions.

Ideally we'd fix the state object, but re-parsing the window location isn't
that inefficient and simplifies code.