Add and remove interaction callbacks as required
Since this only adds and removes stuff from our own internal hook arrays, we can do this as we like without a performance hit. This way we do not fire off lots of useless pause events on interaction  when we have paused already. We also need to disable these hooks when flying with no boxes show, as the hooks are called even if disable_interation=true