homepage: position "start exploring" absolutely #232
We try to float-right the start exploring pill so the text can wrap around
it in a mobile view. Unfortunately the tree has some
aspect-ratio-preserving behaviour that stretches everything and stops the
pill being at the bottom.

Instead, position it absolutely and use the container to make sure the
text clears it.