controller/controller_anim: Don't assume nodes are leaf nodes #231
The extra - means we assume that the largest node is a leaf node, which is
easily not the case. As a result we end up calculating the ancestor node from a
potentially far-flung leaf node, and jump back up and down as part of the
flight.

Fix the conditional, so our flight path is straight again.