sponsorship_duration_days -> sponsorship_ends
It turns out that we are better saving the actual date of expiry (which we set during validation) rather than incrementing the total sponsorship duration. Also that way if we revalidate, we can reset the validation time and it doesn't change the expiry date. We don't currently use sponsorship_duration_days anyway.