sponsor_renew: Copy over details from expired rows #363
When attempting to renew an expired row, we need to create a new
potential reservation. This should have all the details from the old
one, so copy them over.

Use the writable flag to ensure we don't accidentally copy over payment
details, etc. Set writable = False on all the rows we shouldn't copy.