sponsorship: Rework renewal of expired sponsorships #304
The previous attempt won't work, since we copy details
(user_sponsor_name, verified_time) that we want to keep, but will also
mean that we mark the replacement as paid-for prematurely.

Instead, add a prev_reservation_id that references the old node. Then
on only copy details on confirmation. This means that we can copy
details without triggering paid early, and both renewal and extensions
can share the same detail-copying logic.

Also fill in prev_reservation_id for sponsorship extensions.