sponsor_leaf: Open embedded form in new window regardless #468
Unless a direct result of user interaction, Firefox will not let
javascript open a new window. As a result, we can't both validate form
server-side (and show errors in the iframe), and redirect to PayPal
if valid (in a new window).

Instead, remove the validate_form_via_ajax_reload logic, and always
open embedded sponsor forms in a new window. This is a little uglier
for form revalidations, but at least works.