sponsor_pay: Allow retries from spl_waitpay #479
Populate required fields from the unverified state so that payments can be
attempted again if required.

If a reservation is a "unverified waiting for slow payment", emphasise
they may want to try again.

Use this to both populate a "try again" link in the status page, and as
a place to return to if a user cancels the PayPal transaction.