data_repo: Clear images before setting #389
Before setting images, clear existing images from leaves that are part
of the response. So if there are no images in the response, we don't
leave old images around, which may not be available any more.

Remove embedded jQuery from api_wrapper so we can load under a
unit-test, assume it's been loaded globally already.