tree_settings: Make _cbf a derived property #369
Generally _CBF theme variants are separate themes in their own right,
the only place where we care about the difference is in the UI, where
we have a checkbox rather than listing both themes and _CBF variants.

Moving parsing to/from this format down to the UI, so we don't have to
worry about the distinction elsewhere.