controller: Rewrite reset as leap_to, rationalise calls #530
The implementation of controller.reset() is basically the same as
leap_to(), but without the recursion for obvious reasons. Simplify by
actually using leap_to() here.

Given it's just a leap_to, it's now pretty obvious that we don't need
to reset before move when doing an init_move_to(). Tidy up calls to
.reset() when the next step is to leap somewhere else.