position_helper: Implement leaps with perform_actual_fly #530
A leap is the same as a flight without the animation, so if speed is
infinite, just go straight to the destination. This is slightly less
efficient as we reanchor as a flight would, rather than straight to the
destination, but OTOH we get to throw away heaps of
kinda-similar-but-not-quite code.

This also means that leap_to can support into_node, which wasn't part
of the actual_leap codepath.