Tour: Use CSS classes to trigger tourstops #539
Instead of explictly javascripting show/hide etc, use the state CSS
class to trigger show/hide.

Add some default tourstop CSS that can emulate transition_in_visiblity,
and replace instances of this with relevant classes.

Replace on_* exec events with CSS mutation observers, and include a
helper to handle the common code required here.

The only missing functionality now is block-on-transition, which
automatically noticed when a transition resulted in an empty display
and blocked interaction. This doesn't replace it yet.