controller_tour: Handle tour/screensaver in a common controller #244
Add a controller for managing the Tour/Screensaver classes, so the
views just have to ask for a tour by URL via set_screensaver or
tour_start.

Move default setting-finding out of Tour.js, since the controller is a
more natural location for this sort of code and will ease future unit
testing.

This also gives us a place to co-ordinate between tour/screensaver and
make sure only one is active at any one time, so do that.