Tour: allow for starting at points other than 0 #244
Loosen seams so the start point doesn't have to be zero, so this can
thene be exposed via. the controller at a later point.

In the process, fix prev_stop() which was returning stop 0 when there
is no previous stop. This is only apparent when starting at non-zero,
ordinarly the only case when there isn't a previous stop is when the
current stop is zero.