models: Tourstop identifiers, symlink table #562
Instead of having a many-many relationship for tours<->tourstops, make
tourstops themselves be a 1:many (i.e. each tourstop has a "home"
tour). If we genuinely share tourstops, use the tourstop_symlink table,
which can link to another tour.

Also give tourstops an identifier, this gives us an easy hook to upsert
with, and can just be part of the tour definition as it's only unique
within a tour.