src/Tour/HandlerHTMLAudio: Inline audio #535
HTML Audio handler for inline audio; background narration if such a
thing exists is a separate topic.

Also add an audio embedded for wikipedia commons. This probably isn't
the right way to do it long-term, they have their own player to use,
but it's more as a convenient source of audio.