navigation/setup_page: Import UserInterruptError properly #231 #325
"import UserInterruptError from" doesn't import UserInterruptError,
instead it imports default, and calls it UserInterruptError. The
default in this case is the function that adds leap_to to the
controller.

Thus when we were doing instanceof UserInterruptError, we were actually
seeing if it was an instanceof that function.

What we need to do is "import { UserInterruptError }", which means
import UserInterruptError as UserInterruptError. And we need to
actually export it so this works.