tour/TourStop: Don't pause tours on flight-skip #605 #535
The previous backstop wasn't precise enough, when skip is pressed,
the flight is interrupted but we shouldn't be in pause.

Instead, only enable backstop when both the tourstop is still in
TRANSITION_IN (i.e. we think we're still flying) and the tour isn't
paused already.

In an ideal world this will only trigger when doing
"onezoom.tree_state.flying = false" at the console.