tour/TourStop: Convert transitions to use blocks #535
Narration-style videos will potentially want to block a tourstop from
arriving. To be able to do this convert transitions to work with blocks
in the same way as waiting at tourstops.

In both cases, we wait for all blocks to clear, at which point we can
advance() to the next state for the tourstop. Skip calls advance()
directly and ignores stops.

Instead of the incoming tourstop controlling the state of the
previous, both copy any blocks on themselves (e.g. flights) onto the
other tourstop. This means that during a transition either could stop
the transition from ending (e.g. one has a video playing, the other has
a flight continuing), and we'll wait for both.