Tour/TourStop: Wait for flights to actually finish #535
There were 2 problems with this.transition_promise_active:

* A flight is global to the treeviewer, not a tour (let alone a
tourstop). On pressing "prev" it may not realise that the last tourstop
was flying, and tread on each other's toes
* "tree_state.flying = false" isn't enough to act as a lock. It only
has an effect when the timeout for next frame triggers. Stopping a
flight and immediately starting one up again will mean the flight
machinery doesn't realise that it's supposed to be flying somewhere
else, and goes very wrong.

Make the new cancel_flight() return a promise that resolves when the
flight has finished cancelling, i.e. when a frame tick has happened,
realise it's interrupted and thrown UserInterruptedError.

Then, use this to serialise flights in tours, by trying to cancel a
flight before starting a new one. We don't rely on this to cancel the
previous flight, generally this would have been done by next(), but
stops from rushing ahead and starting the new one.

Thanks to this, the bodge of waiting a tick before advance()ing is no
longer required.