tour/TourStop: On skip forward/back, should be tstate.PLAYING #605
A goto_next/prev should always result in the overall tour being in a
tstate.PLAYING, because the tourstops will be playing. Set our state at
relevant points.

We don't need a full user_pause / user_resume(), since that will just
interfere with the processes of going forwards / backwards, but we do
need to ensure that the attached-ness of the interaction hooks matches.

Instead of manually attaching/removing, attach/remove automatically on
state change.