views/tour/tutorial_MD: Don't pop-up unnecessarily
Cancel the timeout if we've moved out of active_wait before we got a
chance to trigger.

Don't do anything if the tour is paused when the timeout runs out
(being paused will mean the user is tinkering with the tree, and
probably didn't want a pop-up in their face).