tour/TourStop: Apply into_node to leap_to
We now always make a final leap when arriving at a tourstop, which
makes it much more noticable that into_node was forgotten for the
leaps. Use the TourStop's value of into_node for both leaps and
flights.

Convert resume() to just use arrive_at_tourstop(), since for
ACTIVE_WAIT it's an equivalent operation, INACTIVE should never happen
but if it did, play() is the correct choice.