src/Tour: Move state classes to an attribute #535
For autoplay to stop videos properly, we need to listen to all state
changes, rather than explicit x->y ones. Making this work, avoiding all
the noise we generate with other classes is tricky.

Move the state to a separate attribute, where arguably it should have
been to start off with.