tour/HandlerVimeo: Support data-ts_autoplay #535
Look for data-ts_autoplay on the iframe node, expecting a list of
tourstop states in which should be autoplaying.

Only block transition if we're in the last of these state lists, i.e.
if ts_autoplay='tsstate-active_wait tsstate-transition_out' we don't
block active-wait any more, but let the video continue and only block
the transition ending.

Add a special mode to tourstop_observer() to report any state-change,
and rework Vimeo to use this. This allows 2 things:

1) We can notice when a tourstop goes from a no-autoplay state to
inactive, and stop the video if the user started it manually.
2) We can notice the transition between autoplay states, and add the
block once we've flipped to the final state

The other handlers will need refactoring similarly.