tour/TourStop; Move leave() to after transition_in change #535 #606
With changes to the state mechanism, we'd trigger events for leaving
the previous stop before the next one became active, resulting in
autoplay videos adding their blocks to an inactive state and being
thrown away.

Move the leave() call until after tsstate-transition_in has happened,
so we don't do this. Ideally the system wouldn't be so brittle that
this is a problem, mind.