sponsorship: Add expiry_hysteresis to reminder triggers #439
To avoid sending dribbles of reminders, use expiry_hysteresis to batch
any that occur in the subsequent 15 days as well as whatever triggered
the reminder email.

Also, include initial_triggers/final_triggers/days_left in the output,
mostly for debugging at this stage but may be useful to include in the
template.