navigation/record: No get_controller in record_url()
Yet again to remove circular dependencies, stop using get_controller()
in record_url(). To do this, pass in the controller every time it is
called.

The key complication here is the flying_finish hook, which now needs to
be set-up somewhere with the controller available. Bung in
controller.js for now.

Arguably what this change is suggesting is navigation/record.js should
actually be controller/record.js, but that's probably too much churn
for today.