OZentry: Coalesce variable names
Use the same variable names as used in the cut map.

Remove window.cut_threshold fallback, since thanks to a variable name
mismatch it didn't do anything. Likewise tree_date.

Remove out-of-date comments (well, 10 000 is still hard-coded, but by
the generator, we use the value it used so that's good enough for
here).