factory: Check flight status when rendering
reanchor_and_dynamic_load_tree is the only use of dynamic_loading, a
more appropriate place to check for tree_state.flying, and saves a
little bit more work we shouldn't be doing.

It also gets rid of a tree_state import in factory, which is a bit of
an odd dependency.