projection/marked_area: Re-implement marked-areas #544
Rip out the previous marked_areas code, and replace with projection/marked_area
that annotates nodes with their required markings as they are
developed.

Connect this to the existing controller methods so common-ancestor
highlighting continues to work transparently.

Restructure config.marked_area_color_map.push a bit to have a easier to
understand array-of-object formats, although it's going to change again
when the full mark algorithm starts existing.