navigation/setup_page: Unpick change_view_type() race condition
change_view_type() fires off async tasks to rebuild the tree, so other
setup may not have worked on the tree it expects. This is most notably
obvious by:-

* /live/?vis=polytomy
* Change view to "Spiral" (i.e. the default visualisation)
* Mangled mess of both visualisations

Sort this by paying attention to the promise that
tree_settings.rebuild_tree() produces, and wait for it to finish before
continuing.

We only need to explicitly rebuild the tree if a vis isn't chosen, so
move this out of OZentry.