util_factory: Mark sample data nodes as detail_fetched #656
We don't need a populated ott_name_map if the nodes have
node._detail_fetched set, if it is then the name is fetched via.
get_attribute().

Having to do so is ugly, but it means we don't get tripped up by other
things expecting detail_fetched to be set.