controller_navigation: Always build tree via. change_view_type #654
If setting 'vis=spiral', change_view_type would assume it needs to do
nothing, as that's the default, and setup_page_by_state would have
assumed that change_view_type() built the tree.

Fix this by always building the tree in change_view_type() on init.
This is a bit weird, but OTOH already knows what tree type needs to be
set.