controller_anim: Details before leaping to search results #656
leap_to() doesn't fetch the details of where it's leaping to. For
zoomout, etc this makes sense. But for init=jump or anim=jump, this is
needless tree-churn.

Extract code to do this from init=jump, use it for anim=jump.

A previous workaround was to poke the search results into data_repo, so
there is at least some detail ready for the tree. Now that isn't
required, remove.