data_repo: Remove ott_name_map #656
Especially on early tree load, it's possible for data_repo.metadata to
have been populated but this.detail_fetched not to have been set. This
was causing get_attribute() to needlessly fail, falling back on
ott_name_map (which isn't restricted).

Always return metadata if we have any, so in combination with the
previous commit, ott_name_map is now entirely redundant.