views/treeviewer: Wait a bit after language change #655
jQuery inserts the DOM it's loaded into the page, then fires
onSuccess(). This is before UIKit has had a chance to notice the
mutations itself, and set itself up.

Wait just long enough that other events fire, then carry on with
setting up the UI. A bit of a cop-out, but there isn't a post-setup
event we can hook to.