modules/pinpoint: Server-side common-ancestor #646
Use a recursive query to find all parents of given OTTs, then find the
deepest one that's common to all inputs. Plug this into
resolve_pinpoint_to_row().

Note that the client will still work it out for itself in preference to
get the server to do it, as it's much cheaper to do it there.

However, it makes no sense when we also need metadata on the ancestor,
e.g. if in Popular Places.