navigation/pinpoint: Don't return potentially-invalid data #646
The "latin_name" part was originally just to go to /API/search_init.
This is no longer required, so we can instead return sciname like the
server would.

We should be returning an object with consistent values. If we know
that the OTT/sciname is wrong, or not sure if so, remove it so
something downstream doesn't take it as gospel.

Rework the highlight tests to work around the changes.