factory/midnode: Cache meta reference #692 #656
Start the process of per-string caches in place of a reference to the
data_repo, using it to simplify the sponsorship properties.

The other cases transform data out of the data_repo, so will need a bit
more thought.

This will mean we don't have to have a dedicated cache for tours once
we get to that point.