factory/data_repo: Store tours lists & expose on midnode #692
Stash everything in previous commit in node/leaf metadata, fetch it in
midnode.

NB: We don't add a column to either metadata array for this, we assign
it to the array as a property. This means we don't eat up storage for
every node/leaf, only the ones with tours.

It may be worth doing similar with sponsors, which are also going to be
quite sparse.