projection/shapes: Collapse shapes & renderers together
There's a 1:1 mapping between the pairs of modules, making the code
much harder to follow. With renderer functions together with their only
use, the code is much easier to follow.

Replace renderer & path_follow instances with regular
render/path_follow instances on the classes. Keep the actual renderers
as separate functions for now, but no reason not to tidy this up after.