controller/navigation: Accessor for get_largest_visible_node
We always need the controller about to use this, since the controller
has the root node. Add it as a method there, so the base case of
supplying controller.root is no longer a problem.

The controller is also accessible to the Tour object, so the utils
version is no longer required.