tour: Add data-focal-area to constrain focal area #547
Tour layouts will be made with a focal area in mind, so they're the
ones that should decide what the focal areas are in a tour.

Ideally it'd be in the CSS for a layout, but an attribute on the
template is a lot closer to the layout definition.

This is also not very fine-grained. The most flexible way of doing
things would be to add a querystring attribute to set this
per-tourstop, but that could be done later if it's really needed.