scss/tour: New tourstop styling #691
Apply new tourstop styling by default to all database tours, using
data-focal-area to shape the focal area to what the CSS expects.

There are 3 layouts:

* portrait: Lower half of screen is tourstop, with a draggable handle
to resize
* small landscape: lower-left is tourstop, with a draggable handle
* large landscape: Tourstop is a left-hand modal

Note we don't bother with a separate view for large portrait, for the
size of the userbase the same view as mobiles is good enough.