models/db: Revert auto is_testing attempt in aa80e2db
is_testing iff running under rocket doesn't work for unit tests. There
we try and work out if it's a production environment and refuse to run.
But there's no way of telling if it's automagic at runtime.

We could do something like ``grunt prod`` generating a config module
that we read in, but there's probably enough in the air at the moment
without adding to the confusion.