tour: Add goto_stop(tourstop_num) #691
For the progress indicator, we'll need to go to a given tourstop.

Add goto_stop() that does the same as back, but for any stop. The
"backward" interaction is what we want in our current case. Eventually
this would need to be re-thought, but this probably needs 'forward' vs
'backward' rethinking.