scss/search_ui: Static search max-height #718
Attempting to eke out the very last pixels for the height of the search
dropdown isn't working, and isn't really necessary anyway as the search
dropdown is always in the same place.

Having a dropdown that's smaller than the available viewport is more
workable than having one that wanders off because it's too big.