navigation/record: Don't record tour-set QS options in URL #739
When there's an active tour, record_url() should only record the tour
state in the URL, not everything that the tour has set.

This means that if qs_opts is used in a tour, the URL will always
include the pre-tour options, even if the page is reloaded / URL is
shared.