tour/handler/Youtube: Don't stop, seek to start #737
Calling player.stopVideo() will also clear any clip set as part of the
video, to restart we'd have to reload the video, clip and all.

Instead, note the position of the video when it does start, and seek
back there whenever we hide the iframe. This means any tour will have
lots of paused videos off-screen, which may not be ideal.