navigation/setup_page: Remove default_init_pinpoint #742
A default_init_pinpoint will be problematic when switching trees.
Whilst we can change the config setting, ideally this is a per-tree
setting which we can revisit if/when we have an organised home for
such things (e.g. tree_startpoints table).

Duck the problem for now, hard-code the default to @_ozid=1, which is
always the root.