controllers/tour: Reset unspecified fields to DB default #776
If a value is unspecified, update_or_insert() will leave alone on
update. We want missing values to go back to their default, as they
would have done if they weren't specified in the first place.

Add the database defaults at the start of merging the tourstop dict, so
this happens. Add some missing defaults for non-null fields so we merge
the right values.