views/life_MD: Disable tours in museum display #810
The tours contain media copyright links we can't remove, and once
followed you're outside the onezoom sandbox.

We either need to pop-up-ize them like we do with wikipedia, or pull an
attribution line in javascript. Either is relatively fiddly.

Disable tours for now. NB: the tours tab is already disabled, so we
just have to hide the controls at the side.