modules/sponsorship: Force copying username on expire-repurchase #645
The default of the username field is '', not None. This means our
default copying machinations may not trigger.

This was a problem in production, resulting in:

  https://github.com/hyanwong/OZtree/commit/548ecbce15f24fc2f89e148dd282efde91953857

This is essentially a forward-port of that commit, without the linting.

This isn't actually an immediate problem though, thanks to:

  https://github.com/OneZoom/OZtree/commit/5f81922c8833a5cd12aa024dc2faf02ecc211599

...which NULLs old records to preserve view counts, meaning it's very
unlikely that we recreate a record on expire-repurchase.

However, the unit test is still valuable, and there's not much harm in
forcing the username field.