controllers/developer: Unit-tests for embed_edit
Test codepaths in embed_edit so we're less likely to end up with
nonfunctional code again.

Poke enough into the env in util.call_controller() for this to work. In
particular we need a real auth object, as that has .settings.mailer.
Force the username there instead.

Note: We don't go as far as testing the e-mail contents, as we don't
have a fake Mailer / SMTP object setup yet, but if a local mailer is
running you get to see the output.