Common function for static data URLs, fetch within onezoom #978
At some point soon we'll add to the list of static data that OneZoom
expects. Instead of adding extra special cases for these files, invent
a static_data_url_func() in server_urls that can generate a correct URL
with a tree version appended for us.

Use this to move all data fetching to within the api_manager, and
remove the now-redundant dynamic_scripts.html. This means:

* The tree requests aren't backed up behind an /API/version.json
request, reducing load latency
* We don't have to pass in the various generated objects when calling
OZentry, simplifying oz setup
* We only deal with the global objects in one place, making it easier
to replace in future with other formats over the wire.

The big loser in all this is minlife, which now has no automatic option
to load local trees. It will now load from the main server by default,
but the "load-local-first" behaviour could be re-implemented with a
fancier static_data_url_func() at a later stage if required.