WIP: position_helper: anchor to nodes that are on-screen *and* reasonably sized
In a balanced tree, the node below "Animals" has a bounding box
encompassing most of the insects. Thanks to this anchoring won't
consider anything lower.

Instead, insist that nodes are on-screen *and* of a reasonable size,
decreasing the considered size to more tightly follow the flight.

This fixes the initial problem, but now the flight from the Elephant
Hawk-Moth to Elephant Seal is very messy. This is likely a problem
elsewhere.